feat: Add Prometheus service & ffmpeg Prometheus streaming metrics (#38)
* feat: Add Prometheus service to docker-compose.dev.yml
feat: ffmpeg metrics (work in progress)

* chore: Refactor MetricsHandler to use exit_status and video_type labels

The code changes in this commit refactor the MetricsHandler class in metrics.py to use the exit_status and video_type labels when incrementing the subprocess_count and streams_count metrics respectively. This improves the clarity and consistency of the metric labels, making it easier to interpret the collected data.

* refactor: Use logging instead of print for exit code in MetricsHandler

* add dependencies and ENV (for logging) to Dockerfile
fixed subprocess status code metrics, added -9 for kill

* added newline to prometheus.yml

* feat: Refactor MetricsHandler to use exit_status and video_type labels

The code changes in this commit refactor the MetricsHandler class in metrics.py to use the exit_status and video_type labels when incrementing the subprocess_count and streams_count metrics respectively. This improves the clarity and consistency of the metric labels, making it easier to interpret the collected data.

* feat: Refactor MetricsHandler to use exit_status and video_type labels

The code changes in this commit refactor the MetricsHandler class in metrics.py to use the exit_status and video_type labels when incrementing the subprocess_count and streams_count metrics respectively. This improves the clarity and consistency of the metric labels, making it easier to interpret the collected data.

* feat: Refactor MetricsHandler to use exit_status and video_type labels

The code changes in this commit refactor the MetricsHandler class in metrics.py to use the exit_status and video_type labels when incrementing the subprocess_count and streams_count metrics respectively. This improves the clarity and consistency of the metric labels, making it easier to interpret the collected data.

* remove redundant amount=1